backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
# openai==1.40.3
pdfplumber==0.11.0
pytesseract==0.3.13
Pillow==10.4.0
groq
